# BE_E_Learning/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Online Learning API",
        "description": "Tài liệu API cho hệ thống học trực tuyến",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/progress": {
            "get": {
                "tags": [
                    "Progress Tracking"
                ],
                "summary": "Danh sách tất cả tiến độ",
                "operationId": "a866fa5a951bb1fb27d7fb25a2e37a10",
                "responses": {
                    "200": {
                        "description": "Danh sách tiến độ"
                    }
                }
            },
            "post": {
                "tags": [
                    "Progress Tracking"
                ],
                "summary": "Tạo tiến độ",
                "operationId": "207157b87ba5715a6e6ac40224c957a6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "course_id",
                                    "lesson_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "course_id": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "lesson_id": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "is_completed": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tiến độ đã tạo"
                    }
                }
            }
        },
        "/api/progress/{id}": {
            "get": {
                "tags": [
                    "Progress Tracking"
                ],
                "summary": "Chi tiết tiến độ",
                "operationId": "60db45d7b010f85cf594a44159df1863",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chi tiết tiến độ"
                    }
                }
            },
            "put": {
                "tags": [
                    "Progress Tracking"
                ],
                "summary": "Cập nhật tiến độ",
                "operationId": "81df76d723457151cef2178a4234ec85",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "is_completed": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tiến độ đã cập nhật"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Progress Tracking"
                ],
                "summary": "Xoá tiến độ",
                "operationId": "703099cfad0a617a31c94631a6f667ca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xoá tiến độ thành công"
                    }
                }
            }
        },
        "/api/progress/user/{userId}/course/{courseId}": {
            "get": {
                "tags": [
                    "Progress Tracking"
                ],
                "summary": "Lấy tiến độ học tập theo user và course",
                "operationId": "4b979b72c2397edadd3620e37f8a975b",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "courseId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách tiến độ học tập"
                    }
                }
            }
        },
        "/api/progress/complete-lesson": {
            "post": {
                "tags": [
                    "Progress Tracking"
                ],
                "summary": "Đánh dấu hoàn thành bài học",
                "operationId": "2e7bac73eeb59db558852e8fddd5f15f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "course_id",
                                    "lesson_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "course_id": {
                                        "type": "integer"
                                    },
                                    "lesson_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Đã đánh dấu hoàn thành"
                    }
                }
            }
        },
        "/api/questions": {
            "get": {
                "tags": [
                    "Question"
                ],
                "summary": "Lấy danh sách tất cả câu hỏi",
                "operationId": "e6318b9018ff4083fa5de51cc42d9b26",
                "responses": {
                    "200": {
                        "description": "Danh sách câu hỏi"
                    }
                }
            },
            "post": {
                "tags": [
                    "Question"
                ],
                "summary": "Tạo câu hỏi mới",
                "operationId": "9dcfd27c2db51c05d96b650b35eb8206",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "quiz_id",
                                    "content",
                                    "order_number"
                                ],
                                "properties": {
                                    "quiz_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "HTML là viết tắt của từ gì?"
                                    },
                                    "order_number": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Câu hỏi đã tạo"
                    }
                }
            }
        },
        "/api/questions/{id}": {
            "get": {
                "tags": [
                    "Question"
                ],
                "summary": "Chi tiết câu hỏi",
                "operationId": "5dff92e576e18c6638a90f86e6ce4c00",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chi tiết câu hỏi"
                    }
                }
            },
            "put": {
                "tags": [
                    "Question"
                ],
                "summary": "Cập nhật câu hỏi",
                "operationId": "3925fba463bf15ab2205432429122aa9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "content": {
                                        "type": "string",
                                        "example": "Câu hỏi cập nhật"
                                    },
                                    "order_number": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Câu hỏi đã cập nhật"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Question"
                ],
                "summary": "Xoá câu hỏi",
                "operationId": "5324d143f621dad7b3c2990f67c3dd52",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Đã xoá câu hỏi"
                    }
                }
            }
        },
        "/api/quizzes/{quizId}/questions": {
            "get": {
                "tags": [
                    "Question"
                ],
                "summary": "Lấy danh sách câu hỏi theo quiz",
                "operationId": "adf3851a165c0a22c44004ead01532a4",
                "parameters": [
                    {
                        "name": "quizId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách câu hỏi theo quiz"
                    }
                }
            }
        },
...
    },
...
}